Did a lot of stuff
Just look at the player test scene. It's a lot. Things that need to be worked on: making sonud clips play on meteor or turret death is not working, maybe adding a different type of turret or wall for variety. Finally, implementing a UI and making the turrets actually cost resources to build